refactor: 수직 어댑터들 패키지 분리
