+++ app/src/main/java/woowacourse/shopping/feature/goods/adapter/vertical/GoodsAdapter.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.feature.goods.adapter
+package woowacourse.shopping.feature.goods.adapter.vertical
 
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
