+++ app/src/main/java/woowacourse/shopping/feature/goods/adapter/vertical/GoodsClickListener.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.feature.goods.adapter
+package woowacourse.shopping.feature.goods.adapter.vertical
 
 import woowacourse.shopping.domain.model.Goods
