+++ app/src/main/java/woowacourse/shopping/feature/goods/adapter/vertical/GoodsViewHolder.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.feature.goods.adapter
+package woowacourse.shopping.feature.goods.adapter.vertical
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
