+++ app/src/main/java/woowacourse/shopping/feature/goods/adapter/vertical/MoreButtonClickListener.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.feature.goods.adapter
+package woowacourse.shopping.feature.goods.adapter.vertical
 
 fun interface MoreButtonClickListener {
     fun onCLickMoreButton()
